Use floats instead of ints and update documentation

+++ test/mouse_test.py
@@ -324,7 +324,7 @@ class MouseModuleTest(MouseTests):
         self.assertIsInstance(pos, tuple)
         self.assertEqual(len(pos), expected_length)
         for value in pos:
-            self.assertIsInstance(value, int)
+            self.assertIsInstance(value, float)
 
     def test_set_pos__invalid_pos(self):
         """Ensures set_pos handles invalid positions correctly."""
